precision

+++ qed_fermion/hmc_sampler_batch.py
@@ -21,8 +21,8 @@ from qed_fermion.utils.coupling_mat3 import initialize_coupling_mat3, initialize
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 print(f"device: {device}")
 
-dtype = torch.float32
-cdtype = torch.complex64
+dtype = torch.float64
+cdtype = torch.complex128
 print(f"dtype: {dtype}")
 print(f"cdtype: {cdtype}")
